Moved params and checks to their own files
