+++ cogs/matchmaking_params.py
@@ -16,4 +16,5 @@ NUMBER_EMOJIS = (
 
 
 WAIT_AFTER_REJECT = 3600
+FRIENDLIES_TIMEOUT = 40
 GGS_ARENA_COUNTDOWN = 30
